ACEfit

Documentation for ACEfit.

ACEfit.Dat

ACEfit.devec_obs

ACEfit.eval_obs

ACEfit.titerate

ACEfit.vec_obs

ACEfit.Dat — Type

Dat: store one configuration (input, e.g., structure, state, ...) that can have multiple observations attached to it. Fields:

config::Any : the structure

configtype::String : Each dat::Dat belongs to a group identified by a string dat.configtype to allow filtering, and grouping.

obs::Vector{Any} : list of observations

meta::Dict{String, Any} : any additional meta information that we may want to attach to this data point; this needs to be raw json.

ACEfit.devec_obs — Function

convert a Vector{T} to some real data, e.g.,

x::Vector{Float64}
devec_obs(::Type{ObsVirial}, x) = [ x[1] x[2] x[3]; 
                                    x[2] x[4] x[5];
                                    x[3] x[5] x[6] ]

ACEfit.eval_obs — Function

Evaluate a specific observation type: Given an observation obs, a model model and a configuration cfg = dat.config, the call

eval_obs(obs, model, cfg)

must return the corresponding observation. For example, if obs::ObsPotentialEnergy and cfg = at::Atoms, and model is an interatomic potential, then

eval_obs(obs::ObsPotentialEnergy, model, cfg) = 
      ObsPotentialEnergy( energy(model, cfg) )

ACEfit.titerate — Method

titerate(f, data; kwargs...) Multi-threaded map loop. At each iteration the function f(dat) is executed, where dat in data. The order is not necessarily preserved. In fact the costs array is used to sort the data by decreasing cost to ensure the most costly configurations are encountered first. This helps avoid threads without work at the end of the loop.

ACEfit.vec_obs — Function

convert some real data, in some generic format, into a vector to be stored in a Dat or Lsq system. E.g.,

V = virial(...)::Matrix{Float64}
obsV = ObsVirial(V)
vec_obs(obsV::ObsVirial) = obsV.V[ [1,2,3,5,6,9] ]  # NB: V is symmetric