ACEfit

Documentation for ACEfit.

Scikit-learn solvers

To use Python based Scikit-learn solvers you need to load PythonCall in addition to ACEfit.

using ACEfit
using PythonCall

ACEfit.AbstractData

ACEfit.BLR

ACEfit.LSQR

ACEfit.QR

ACEfit.RRQR

ACEfit.SKLEARN_ARD

ACEfit.SKLEARN_BRR

ACEfit.assemble

ACEfit.count_observations

ACEfit.feature_matrix

ACEfit.target_vector

ACEfit.weight_vector

ACEfit.AbstractData — Type

ACEfit users should define a type of the form: UserData <: AbstractData

Several functions acting on such a type should be implemented: countobservations featurematrix targetvector weightvector

ACEfit.BLR — Type

struct BLR : Bayesian linear regression

Refer to bayesianlinear.jl (for now) for kwarg definitions.

ACEfit.LSQR — Type

LSQR

ACEfit.QR — Type

struct QR : linear least squares solver, using standard QR factorisation; this solver computes

\[ θ = \arg\min \| A \theta - y \|^2 + \lambda \| P \theta \|^2\]

Constructor

ACEfit.QR(; lambda = 0.0, P = nothing)

where

λ : regularisation parameter

P : right-preconditioner / tychonov operator

ACEfit.RRQR — Type

struct RRQR : linear least squares solver, using rank-revealing QR factorisation, which can sometimes be more robust / faster than the standard regularised QR factorisation. This solver first transforms the parameters $\theta_P = P \theta$, then solves

\[ θ = \arg\min \| A P^{-1} \theta_P - y \|^2\]

where the truncation tolerance is given by the rtol parameter, and finally reverses the transformation. This uses the pqrfact of LowRankApprox.jl; For further details see the documentation of LowRankApprox.jl.

Crucially, note that this algorithm is not deterministic; the results can change slightly between applications.

Constructor

ACEfit.RRQR(; rtol = 1e-15, P = I)

where

rtol : truncation tolerance

P : right-preconditioner / tychonov operator

ACEfit.SKLEARN_ARD — Type

SKLEARN_ARD

ACEfit.SKLEARN_BRR — Type

SKLEARN_BRR

ACEfit.assemble — Method

Assemble feature matrix, target vector, and weight vector for given data and basis.

ACEfit.count_observations — Method

Returns the corresponding number of rows in the design matrix.

ACEfit.feature_matrix — Method

Returns the corresponding design matrix (A) entries.

ACEfit.target_vector — Method

Returns the corresponding target vector (Y) entries.

ACEfit.weight_vector — Method

Returns the corresponding weight vector (W) entries.