ACEfit

Documentation for ACEfit.

ACEfit.BARD

ACEfit.BL

ACEfit.Dat

ACEfit.LSQR

ACEfit.QR

ACEfit.RRQR

ACEfit.SKLEARN_ARD

ACEfit.SKLEARN_BRR

ACEfit.basis_obs

ACEfit.create_solver

ACEfit.devec_obs

ACEfit.eval_obs

ACEfit.set_weights!

ACEfit.titerate

ACEfit.vec_obs

ACEfit.BARD — Type

Bayesian ARD

ACEfit.BL — Type

Bayesian Linear

ACEfit.Dat — Type

Dat: store one configuration (input, e.g., structure, state, ...) that can have multiple observations attached to it. Fields:

config::Any : the structure

configtype::String : Each dat::Dat belongs to a group identified by a string dat.configtype to allow filtering, and grouping.

obs::Vector{Any} : list of observations

meta::Dict{String, Any} : any additional meta information that we may want to attach to this data point; this needs to be raw json.

ACEfit.LSQR — Type

LSQR

ACEfit.QR — Type

struct QR : linear least squares solver, using standard QR factorisation; this solver computes

\[ θ = \arg\min \| A \theta - y \|^2 + \lambda \| P \theta \|^2\]

Constructor

ACEfit.QR(; λ = 0.0, P = nothing)

where

λ : regularisation parameter

P : right-preconditioner / tychonov operator

ACEfit.RRQR — Type

struct RRQR : linear least squares solver, using rank-revealing QR factorisation, which can sometimes be more robust / faster than the standard regularised QR factorisation. This solver first transforms the parameters $\theta_P = P \theta$, then solves

\[ θ = \arg\min \| A P^{-1} \theta_P - y \|^2\]

where the truncation tolerance is given by the rtol parameter, and finally reverses the transformation. This uses the pqrfact of LowRankApprox.jl; For further details see the documentation of LowRankApprox.jl.

Crucially, note that this algorithm is not deterministic; the results can change slightly between applications.

Constructor

ACEfit.RRQR(; rtol = 1e-15, P = I)

where

rtol : truncation tolerance

P : right-preconditioner / tychonov operator

ACEfit.SKLEARN_ARD — Type

SKLEARN_ARD

ACEfit.SKLEARN_BRR — Type

SKLEARN_BRR

ACEfit.basis_obs — Function

function basis_obs end - evaluate an observation on a basis and a config; used for llsq assembly,

basis_obs(obstype, basis, cfg)

For examle, for a potential energy the implementatin might look like this:

basis_obs(::Type{TOBS}, basis, at) where {TOBS <: ObsPotentialEnergy} = 
   TOBS.( energy(basis, at) )

Note how the TOBS operation is now broadcast since energy(basis, at) returns a vector of observations (one for each basis functions) whereas for a model we would use eval_obs.

ACEfit.create_solver — Method

create_solver

ACEfit.devec_obs — Function

convert a Vector{T} to some real data, e.g.,

x::Vector{Float64}
devec_obs(::Type{ObsVirial}, x) = [ x[1] x[2] x[3]; 
                                    x[2] x[4] x[5];
                                    x[3] x[5] x[6] ]

ACEfit.eval_obs — Function

Evaluate a specific observation type: Given an observation obs, a model model and a configuration cfg = dat.config, the call

eval_obs(obstype, model, cfg)

must return the corresponding observation. For example, if obs::ObsPotentialEnergy and cfg = at::Atoms, and model is an interatomic potential, then

eval_obs(::Type{TOBS}, model, cfg) where {TOBS <: ObsPotentialEnergy} = 
      TOBS( energy(model, cfg) )

ACEfit.set_weights! — Function

TODO: needs documentation

ACEfit.titerate — Method

titerate(f, data; kwargs...) Multi-threaded map loop. At each iteration the function f(dat) is executed, where dat in data. The order is not necessarily preserved. In fact the costs array is used to sort the data by decreasing cost to ensure the most costly configurations are encountered first. This helps avoid threads without work at the end of the loop.

ACEfit.vec_obs — Function

convert some real data, in some generic format, into a vector to be stored in a Dat or Lsq system. E.g.,

V = virial(...)::Matrix{Float64}
obsV = ObsVirial(V)
vec_obs(obsV::ObsVirial) = obsV.V[ [1,2,3,5,6,9] ]  # NB: V is symmetric