ACEfit

Documentation for ACEfit.

ACEfit.BARD

ACEfit.BL

ACEfit.BayesianLinearRegressionSVD

ACEfit.LSQR

ACEfit.QR

ACEfit.RRQR

ACEfit.SKLEARN_ARD

ACEfit.SKLEARN_BRR

ACEfit.create_solver

ACEfit.BARD — Type

Bayesian ARD

ACEfit.BL — Type

Bayesian Linear Regression

ACEfit.BayesianLinearRegressionSVD — Type

Bayesian Linear Regression SVD

ACEfit.LSQR — Type

LSQR

ACEfit.QR — Type

struct QR : linear least squares solver, using standard QR factorisation; this solver computes

\[ θ = \arg\min \| A \theta - y \|^2 + \lambda \| P \theta \|^2\]

Constructor

ACEfit.QR(; lambda = 0.0, P = nothing)

where

λ : regularisation parameter

P : right-preconditioner / tychonov operator

ACEfit.RRQR — Type

struct RRQR : linear least squares solver, using rank-revealing QR factorisation, which can sometimes be more robust / faster than the standard regularised QR factorisation. This solver first transforms the parameters $\theta_P = P \theta$, then solves

\[ θ = \arg\min \| A P^{-1} \theta_P - y \|^2\]

where the truncation tolerance is given by the rtol parameter, and finally reverses the transformation. This uses the pqrfact of LowRankApprox.jl; For further details see the documentation of LowRankApprox.jl.

Crucially, note that this algorithm is not deterministic; the results can change slightly between applications.

Constructor

ACEfit.RRQR(; rtol = 1e-15, P = I)

where

rtol : truncation tolerance

P : right-preconditioner / tychonov operator

ACEfit.SKLEARN_ARD — Type

SKLEARN_ARD

ACEfit.SKLEARN_BRR — Type

SKLEARN_BRR

ACEfit.create_solver — Method

create_solver(params::Dict)

Convenience function for creating a solver. The params should contain a type, whose value is a solver type. The remaining params are passed as keyword arguments to the solver's constructor.

Valid solver types: "QR, LSQR, RRQR, SKLEARNBRR, SKLEARNARD"